% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_filtering_gc_data.R
\name{add_linshifts2}
\alias{add_linshifts2}
\title{Internal functions used for filtering steps in the processing}
\usage{
add_linshifts2(dx = NULL, rt_col_name = NULL, Logbook = NULL)
}
\description{
Internal functions used for filtering steps in the processing
}
\keyword{internal}
