% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_individual_peaks.R
\name{align_individual_peaks}
\alias{align_individual_peaks}
\title{algorithm to adjust unreliability of chromatograms on an individual peak basis}
\usage{
align_individual_peaks(chromatograms, error_span = 0.02, n_iter = 1)
}
\arguments{
\item{chromatograms}{\code{data.frame} containing GC data (retention times, peak area, peak hight etc) for
one individual in adjacent columns. The first column for all individuals has to be the retention
time, retention time has to be named RT.}

\item{error_span}{maximum error under which two retention times are still counted as the same peak}

\item{n_iter}{number of iterations for the algorithm. One iteration is usually optimal, but
for large datasets more iteration might be better. Has to be evaluated properly.}
}
\value{
aligned chromatograms
}
\description{
algorithm to adjust unreliability of chromatograms on an individual peak basis
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

