% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_filtering_gc_data.R
\name{align_var}
\alias{align_var}
\title{Clean-Up and formatting of Peak lists}
\usage{
align_var(gc_peak_list, rt_col_name)
}
\description{
Set of functions used internally to to format the data for efficient processing.
Steps include the removal of redundant rows (i.e. no sample carries a substance)
}
\keyword{internal}
