% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_transformation.R
\name{best_shift}
\alias{best_shift}
\title{Selects the optimal Shifting time leading to the maximum number of shared peaks}
\usage{
best_shift(list)
}
\arguments{
\item{list, }{output from shared_peaks}
}
\value{
list of number of peaks and shift step for which this is maximised
}
\description{
Selects the optimal Shifting time leading to the maximum number of shared peaks
}
\details{
If competing shifts exists (i.e. same number of peaks are shared),
         selects the smallest absolute value of a shift
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
  (meinolf.ottensmann@web.de)
}

