% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfla.R
\docType{data}
\name{bfla}
\alias{bfla}
\title{Peak data from a chemical dataset of \emph{Bombus flavifrons}}
\format{A table in the input format of GCalignR, containing peak data ("RT", "Area") of 11 samples.}
\source{
http://onlinelibrary.wiley.com/store/10.1002/jssc.201300388/asset/supinfo/jssc3437-sup-0001-TableS1.zip?v=1&s=57d5d58273d1d4207e70c72cecd5bba4b1fe95a1
}
\description{
This dataset is comprised of chromatography data obtained from cephalic labial gland secretion of the bumblebee \emph{Bombus flavifrons} (Dellicour and Lecocq 2013).
}
\references{
Dellicour, S. and Lecocq, T. (2013): GCALIGNER 1.0: An alignment program to compute a multiple sample comparison data matrix from large eco-chemical datasets obtained by GC. J. Sep. Science 36. DOI:10.1002/jssc.201300388
}
\keyword{datasets}

