% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Check data input}
\usage{
check_data(data, sep = "\\t")
}
\arguments{
\item{data}{path to data file or list of data.frames}

\item{sep}{The field separator character. Values on each line of the file are separated by this
character. The default is tab seperated (sep = '\\t'). See \code{sep} argument in \code{\link[utils]{read.table}} for details.}
}
\value{
TRUE if data is formatted correctly, warning and explanation if not.


 @author Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
 (meinolf.ottensmann@web.de)
}
\description{
Checks whether the data is formatted correctly. The data can either be a .txt file
or a list of data.frames. See \code{\link{align_chromatograms}}
}
\examples{
data(gc_peak_data)
gc_peak_data <- gc_peak_data[1:4]
check_data(gc_peak_data)

}

