% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rows.R
\name{check_redundancy}
\alias{check_redundancy}
\title{Check whether similar rows are redundant or not}
\usage{
check_redundancy(chromatogram, similar, rt_col_name)
}
\arguments{
\item{chromatogram}{\code{data.frame} containing GC data}

\item{similar}{\code{integer}, position of a row, that is similar to the previous}

\item{rt_col_name}{\code{character}, name of the retention time column.}
}
\value{
\item{redundant}{binary \code{vector} coding redundancy by 1}
}
\description{
This function evaluates if rows of similar retention times are redundant.
             This is as long as any sample contains peaks in either one or none of the
             row-pair, but never in both.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

