% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_gc_mat_to_list.R
\name{conv_gc_mat_to_list}
\alias{conv_gc_mat_to_list}
\title{transform GC data to list of individual GC matrices}
\usage{
conv_gc_mat_to_list(all_gc_mat, ind_names, var_names)
}
\arguments{
\item{all_gc_mat}{matrix or data.frame containing GC data (retention times, peak area, peak hight etc) for all individuals
in adjacent columns. The first column for all individuals has to be the retention time}

\item{ind_names}{character vector with names of individuals}

\item{var_names}{character vector with names of columns from all_gc_mat,
such as c("RT", "Area", "Height"). The first name has to specify the retention time}
}
\value{
List of data.frames. Each list element is the GC data for one individual
}
\description{
Transforms a Matrix containing GC data (retention times, peak area, peak hight etc) in adjacent
columns.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}
\references{

}

