% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_filtering_gc_data.R
\name{delete_empty_rows}
\alias{delete_empty_rows}
\title{Clean-Up and formatting of Peak lists}
\usage{
delete_empty_rows(gc_peak_df, average_rts)
}
\description{
Set of functions used internally to to format the data for efficient processing.
Steps include the removal of redundant rows (i.e. no sample carries a substance)
}
\keyword{internal}

