% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_space_colnames.R
\name{delete_space_colnames}
\alias{delete_space_colnames}
\title{Remove spaces in column names}
\usage{
delete_space_colnames(data)
}
\arguments{
\item{data}{data.frame containing GC data.}
}
\value{
data.frame containing GC data with removed spaces in column names
}
\description{
If present, spaces within column names are eliminated
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}
\references{

}

