% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fun_Fact.R
\name{Fun_Fact}
\alias{Fun_Fact}
\title{Funfacts about birds and maybee other wildlife}
\usage{
Fun_Fact()
}
\value{
\item{text}{a short trivia}
}
\description{
randomly selects a trivia from the list  of astonishing facts
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
 (meinolf.ottensmann@web.de)

 @export

 @
}
\references{
https://www.audubon.org/news/11-fun-facts-about-owls
 https://www.audubon.org/news/9-fun-facts-about-turkeys
 https://www.audubon.org/news/9-awesome-facts-about-bird-migration
}

