% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_peak_data.R
\name{gc_peak_data}
\alias{gc_peak_data}
\title{Gas-chromatography data from from three plover species in Madagascar.}
\format{A  \code{list} of \code{data.frame}s. Each \code{data.frame} contains gas-chromatography
peak data for an individual. The variables within each \code{data.frame} are:
 "RT" (retention time of a peak), "start_RT" (the
start of the peak in the chromatogram), "end_RT" (the end of the peak in the chromatogram),
"area" (the area under the peak), "rel_area" (the relative area under the peak), "height" (the
height of the peak), "rel_height" (the relative height of the peak). Each \code{list} element,
i.e. each \code{data.frame} is named with the individual ID.}
\description{
Gas-chromatography data from from three plover species in Madagascar.
}
\references{
Ottensmann, M., Stoffel, M.A., Eberhardt-Phillips, L., in preparation
}
\keyword{datasets}

