% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_peaks.R
\name{gc_peaks}
\alias{gc_peaks}
\title{Gas-chromatography data of three \emph{Charadriidae} species from Madagascar}
\format{A  list of data frames. Each data frame contains gas-chromatography peak data of an sample.
The variables within each \code{data.frame} are: "time" (peak retention time) and "area" (integral of the peak curve).
Each list element i.e. each data frame is named with the unique sample ID.}
\description{
This is in example of a typical gas-chromatography output file, listing a number of peaks with their
respective retention times and abundance measures. Peaks were detected using Xcalibur 2.0.5 (Thermo Fisher Scientific)
}
\references{
Ottensmann, M., Stoffel, M.A., Eberhardt-Phillips, L., in preparation
}
\keyword{datasets}

