% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_peaks_aligned.R
\name{gc_peaks_aligned}
\alias{gc_peaks_aligned}
\title{Aligned Gas-chromatography data of three \emph{Charadriidae} species from Madagascar}
\format{Object of class "GCalign" including three lists. List \strong{"aligned"} includes data.frames
for all variables present in the raw data ("time" and "area"). The list \strong{"heatmap"} holds data frames
with retention times of the input data, linearly adjusted retention times as well as the final output, were
peaks are aligned among samples. This file is primarily used in \code{\link{gc_heatmap}}. The list \strong{"Logfile"}
summarises the alignment process and the data structure before, during and after running \code{\link{align_chromatograms}}.
For a convenient overview use \code{\link{summary.GCalign}}.}
\description{
This is in example of an aligned gas-chromatography dataset processed with \code{\link{align_chromatograms}}.
The raw data is accessible within this package \strong{gc_peaks.RData}
}
\references{
Ottensmann, M., Stoffel, M.A., Eberhardt-Phillips, L., in preparation
}
\keyword{datasets}

