% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_transformation.R
\name{linear_transformation}
\alias{linear_transformation}
\title{maximisation of the number of shared peaks with reference chromatogram}
\usage{
linear_transformation(chromatograms, reference, shift = 0.05,
  step_size = 0.01, error = 0, rt_col_name)
}
\arguments{
\item{chromatograms}{list of data.frames with each data.frame being an individuals gc data}

\item{rt_col_name}{character string for name of retention time in gc table}

\item{high}{Upper threshold for retention times. RTs lower than \code{high} will be kept}
}
\value{
gc table with retention times cut
}
\description{
maximisation of the number of shared peaks with reference chromatogram
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}
\references{

}

