% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_append.R
\name{matrix_append}
\alias{matrix_append}
\title{Append zeros to matrices}
\usage{
matrix_append(matrix, chromatograms)
}
\arguments{
\item{matrix}{a matrix to append.}

\item{chromatograms}{list of matrices.}
}
\value{
a appended matrix.
}
\description{
\code{matrix_append()} adds zeros to a matrix or a data.frames grouped within a list.
If required rows containing just zeros are appended to match the dimensions of the
largest matrix having the most rows of the list \code{chromatograms}.
}
\details{
If the matrix has the same dimensions as the largest of \code{chromatograms}
         the input matrix is returned.
}

