% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_chroma.R
\name{mean_per_row}
\alias{mean_per_row}
\title{calculates mean retention time per row}
\usage{
mean_per_row(chromatograms, rt_col_name)
}
\arguments{
\item{chromatograms}{data.frame containing GC data (retention times, peak area, peak hight etc) for
  one individual in adjacent columns. The first column for all individuals has to be the retention
  time, retention time has to be named RT.

  @param rt_col_name character string denoting the column containing retention times in \code{chromatograms}}
}
\value{
vector with mean retention times per row in \code{chromatograms}.
}
\description{
calculates mean retention time per row
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

