% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_retention_time_row.R
\name{mean_retention_time_row}
\alias{mean_retention_time_row}
\title{Calculate mean retention for one peak}
\usage{
mean_retention_time_row(gc_peak_list, samples, retention_row, rt_col_name)
}
\arguments{
\item{gc_peak_list}{List of data.frames. Each data.frame contains GC-data (e.g. retention time, peak area, peak height) of one sample. Variables are stored in columns.
Rows represent distinct peaks. Retention time is a required variable.}

\item{samples}{indices of samples to consider in the calculation of the mean retention time.}

\item{retention_row}{index of the current retention time to be evaluate.}

\item{rt_col_name}{Character string naming the column holding retention times.The designated variable needs to be numeric.}
}
\value{
mean retention time for peaks at index \code{retention_row}.
}
\description{
\code{mean_retention_time_row} calculates the average retention time of a row. Samples in \code{gc_peak_list}
lacking a peak (i.e retention time = 0) are not considered.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

