% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_redundant_peaks.R
\name{merge_redundant_peaks}
\alias{merge_redundant_peaks}
\title{Merges redundant rows}
\usage{
merge_redundant_peaks(gc_peak_list, min_diff_peak2peak = 0.05, rt_col_name,
  conc_col_name, criterion = "strict")
}
\arguments{
\item{gc_peak_list}{List of data.frames. Each data.frame contains GC-data (e.g. retention time, peak area, peak height) of one sample. Variables are stored in columns.
Rows represent distinct peaks. Retention time is a required variable.}

\item{min_diff_peak2peak}{Defines the desired minimum difference in retention times among distinct
substances. Substances that differ less, are merged if every sample contains either one
or none of the respective compounds. This parameter is a major determinant in the classification
of distinct peaks. Therefore careful consideration is required to adjust this setting to
your needs (i.e. the resolution of your gas-chromatography pipeline). Large values might cause the
merge of true peaks, if those are not occuring within one individual, which might happen by chance
with a low sample size. Small values can be considered as conservative. Default is 0.02.}

\item{rt_col_name}{Character string naming the column holding retention times.The designated variable needs to be numeric.}

\item{conc_col_name}{Character string naming a column used for quantification of peaks (e.g. peak area or peak height).
The designated variable needs to be numeric.}

\item{criterion}{character string either "strict" or "proportional". "strict" is used to delete the row containing zeros,
"proportional" provokes the selection of the larger of two peaks, defined by the concentration of the peaks.
\code{criterion="proportional"} will only affect pairs of peaks where less than 5% of the samples containg
two conflicting peaks.}
}
\value{
a list of data.frames for individual samples with redundant peaks merged.
}
\description{
\code{merge_redundant_peaks()} allows to merge neighbouring peaks (i.e substances) of similiar retention time. All peaks
that are similar in retention times (definded by \code{min_diff_peak2peak}) are combined, in case that
samples contain either one or none of two compared peaks, but never both. An optional procedure
allows to merge close peaks if only a small subset of samples contains both of them.
}
\details{
Peaks that are potentially redundant are selected by comparing the similarity of retention times with
respect to \code{min_diff_peak2peak}. Candidates are peaks that are closer to each other than this
threshold. After one pair of rows was merged, average retention times are calculated again. The algorithm
stops if no pair of similar peaks remains or none of those appears redundant (i.e. at least one sample
shows both peaks). As a last step, peaks that were not proved to be redundant by this strict criterion,
are merged if only a small proportion of samples (less than 5%) shows to conflicting peaks. In this
case the peak with the larger concentration (peak area or peak height, depends on \code{conc_col_name})
is retained and the other one deleted.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

