% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rows.R
\name{merge_rows}
\alias{merge_rows}
\title{Merge two rows into one}
\usage{
merge_rows(gc_peak_df, to_merge, criterion = "zero", rt_col_name)
}
\arguments{
\item{gc_peak_df}{data.frame containing GC-data (e.g. retention time, peak area, peak height) of one sample. Variables are stored in columns.
Rows represent distinct peaks. Retention time is a required variable.}

\item{to_merge}{integer vector indicating the index of the last row of a pair to merge.}

\item{criterion}{character string either "zero" or "area". "zero" is used to delete the row containing zeros,
"area" provokes the selection of the larger of two peaks, defined by the concentration of the peaks.
"area" is only useful when \code{criterion} is "proportional".}

\item{rt_col_name}{\code{character}, name of the column holding retention times (i.e. "RT")}
}
\value{
aligned chromatograms
}
\description{
Merges the content of two rows in such a way that content is retained. This means that either the row
is selected that contains in case that the second is empty (i.e just zeros). If both rows have a content
the row with the larger concentration (i.e peak area or peak height) is retained in \code{gc_peak_df}
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

