% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_transformation.R
\name{peak_shift}
\alias{peak_shift}
\title{shifts retention times of a chromatogram and estimates the number of shared peaks with the
reference.}
\usage{
peak_shift(sample_df, ref_df, shift = 0.05, step_size = 0.01, error = 0,
  rt_col_name)
}
\arguments{
\item{sample_df}{\code{data.frame} with individual gc data}

\item{ref_df}{\code{data.frame} with individual gc data of the reference chromatogram}

\item{shift}{shift span to try for maximising shared peaks}

\item{step_size}{steps to take within the shift span}

\item{error}{allowed error}

\item{rt_col_name}{RT time column name}
}
\value{
Numeric Value, indicating the best shift (e.g. -0.02 seconds)
}
\description{
shifts retention times of a chromatogram and estimates the number of shared peaks with the
reference.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
  (meinolf.ottensmann@web.de)
}

