% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_best_reference.R
\name{pick_best_reference}
\alias{pick_best_reference}
\title{Picks the most suitable Reference for systematic shift corrections}
\usage{
pick_best_reference(gc_peak_list = NULL, rt_col_name = NULL)
}
\arguments{
\item{gc_peak_list}{List of individual samples, where samples a data frames of numerical variables in columns. A column needs to contain retention times of peaks.}

\item{rt_col_name}{Name of the column in data frames of \code{gc_peak_list} that contains retention times.}
}
\description{
Selects the chromatogram that shows the highest average similarity to all other samples a the reference for linear corrections of systematic shifts in peak retention times. The determination is grounded on the median number of shared peaks between a candidate reference chromatogram and all others. The function compares evaluates the suitability of every sample supplied.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

