% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.GCalign.R
\name{plot.GCalign}
\alias{plot.GCalign}
\title{plot.GCalign}
\usage{
\method{plot}{GCalign}(GcOut)
}
\description{
plot.GCalign creates basic plots visualising the alignment accuracy of chromatograms that
have been processed by \code{\link{align_chromatograms}}
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
 (meinolf.ottensmann@web.de)

 @export
}

