% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_empower2.R
\name{read_empower2}
\alias{read_empower2}
\title{Import data from single EMPOWER2 HPLC files}
\usage{
read_empower2(
  path = NULL,
  pattern = ".txt",
  sep = "\\t",
  skip = 2,
  id = "SampleName"
)
}
\arguments{
\item{path}{path to a folder containing input files}

\item{pattern}{pattern used to select files. By default ".txt"}

\item{sep}{The field separator character. The default is tab separated (\code{sep = '\\t'}).
See the "sep" argument in \code{\link[utils]{read.table}} for details.}

\item{skip}{rows to skip before reading data}

\item{id}{column containing sample name}
}
\value{
a list of data frames (each corresponding to a sample)
}
\description{
reads output files of the EMPOWER 2 SOFTWARE (Waters). Input files must contain data of single samples deposited within the same directory.
}
\keyword{beta}
