% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_peak_list.R
\name{read_peak_list}
\alias{read_peak_list}
\title{Read content of a text file}
\usage{
read_peak_list(data, sep = "\\t", rt_col_name)
}
\arguments{
\item{data}{a text file}

\item{sep}{The field separator character. The default is tab separated (\code{sep = '\\t'}).
See the "sep" argument in \code{\link[utils]{read.table}} for details.}

\item{rt_col_name}{Character string - the name of the column containing the retention times.The variable needs to
be numeric and the decimal separator needs to be a point.}
}
\value{
a list of peak lists for each sample
}
\description{
reads the content of text file that is formatted as described in \code{\link{align_chromatograms}} and creates a peak list.
}
