% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_cols.R
\name{rename_cols}
\alias{rename_cols}
\title{Renames table columns in GC list}
\usage{
rename_cols(data, var_names)
}
\arguments{
\item{data}{\code{data.frame} containing GC data (retention times, peak area, peak hight etc) for
one individual in adjacent columns. The first column for all individuals has to be the retention
time}

\item{var_names}{character vector with names of columns for \code{data}, such as c("RT", "Area",
"Height"). The first name has to specify the retention time}
}
\value{
\code{data.frame} with renamed columns
}
\description{
Renames table columns in GC list
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
  (meinolf.ottensmann@web.de)
}
\references{

}

