% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_cutoff.R
\name{rt_cutoff}
\alias{rt_cutoff}
\title{cuts retention time below and above specified times}
\usage{
rt_cutoff(data, rt_col_name, low = NULL, high = NULL)
}
\arguments{
\item{data}{\code{data.frame} containing GC data (retention times, peak area, peak hight etc) for
one individual in adjacent columns. The first column for all individuals has to be the retention
time}

\item{rt_col_name}{character string for name of retention time in gc table}

\item{low}{Lower threshold for retention times. RTs higher than \code{low} will be kept}

\item{high}{Upper threshold for retention times. RTs lower than \code{high} will be kept}
}
\value{
gc table with retention times cut NAs cut
}
\description{
cuts retention time below and above specified times
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}
\references{

}

