% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_extract.R
\name{RT_extract}
\alias{RT_extract}
\title{RT_extract extracts the retention times of all objects of a list
 of chromatograms and writes them to a matrix, where rows are samples
 and columns retention times. Columnames represent the order of substances
 starting with the lowest retention time}
\usage{
RT_extract(chromatograms)
}
\arguments{
\item{chromatograms}{List of Chromatograms with just a column for Retention Times}
}
\value{

}
\description{
RT_extract extracts the retention times of all objects of a list
 of chromatograms and writes them to a matrix, where rows are samples
 and columns retention times. Columnames represent the order of substances
 starting with the lowest retention time
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}
\references{

}

