% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_extract_heatmap.R
\name{rt_extract_heatmap}
\alias{rt_extract_heatmap}
\title{rt_extract_heatmap}
\usage{
rt_extract_heatmap(chromatograms, blanks, rt_name, del_single_sub)
}
\arguments{
\item{chromatograms}{List of chromatograms}

\item{blanks}{character vector of blanks}

\item{rt_name}{chracter, specifies column holding retention times}

\item{del_single_sub}{logical, indicating wheter single substances are deleted}
}
\value{

}
\description{
rt_extract_heatmap
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}
\references{

}

