% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seal_factors.R
\docType{data}
\name{seal_factors}
\alias{seal_factors}
\title{Factors corresponding to gas-chromatography data of Antarctic Fur Seals (\emph{Arctocephalus gazella})}
\format{A data frame where columns represent factors, rows are samples.}
\source{
http://www.pnas.org/content/suppl/2015/08/05/1506076112.DCSupplemental/pnas.1506076112.sd02.xlsx
}
\description{
List of factors corresponding to samples in \code{\link{seal_peaks}}
}
\references{
Stoffel, M.A.; Caspers, B.A.; Forcada, J.; Giannakara, A.; Baier, M.; Eberhart-Phillips, L.; Müller, C.; Hoffman, J.I. (2015): Chemical fingerprints encode mother-offspring similarity, colony membership, relatedness, and genetic quality in fur seals. In: Proceedings of the National Academy of Sciences of the United States of America 112 (36), S. E5005-12. DOI: 10.1073/pnas.1506076112.
}
\keyword{datasets}

