% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_rows.R
\name{shift_rows}
\alias{shift_rows}
\title{inserts row of zeros in chromatogram}
\usage{
shift_rows(chromatograms, current_sample_index, retention_row)
}
\arguments{
\item{chromatograms}{\code{data.frame} containing GC data (retention times, peak area, peak hight etc) for
one individual in adjacent columns. The first column for all individuals has to be the retention
time, retention time has to be named RT.}

\item{current_sample_index}{index of sample to insert zeros}

\item{retention_row}{current retention time row in which to insert zeros}
}
\value{
chromatogram \code{data.frame} with inserted zeros in a given row
}
\description{
inserts row of zeros in chromatogram
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

