% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar_peaks.R
\name{similar_peaks}
\alias{similar_peaks}
\title{Find peaks with similar retention time}
\usage{
similar_peaks(average_rts, min_diff_peak2peak = 0.05)
}
\arguments{
\item{average_rts}{vector of average retention times per peak across individuals.}

\item{min_diff_peak2peak}{Defines the minimum difference in retention times among distinct
substances. Substances that differ less, are merged if every sample contains either one
or none of the respective compounds. This parameter is a major determinant in the classification
of distinct peaks. Therefore careful consideration is required to adjust this setting to
your needs (i.e. the resolution of your gas-chromatography pipeline). Large values might cause the
merge of true peaks, if those are not occuring within one individual, which might happen by chance
with a low sample size. Small values can be considered as conservative. Default is 0.02.}
}
\value{
integer vector of indices pointing to the last index of a row-pair that is similar
}
\description{
Estimate degree of similarity between peaks by comparison of mean retention times of across samples.
}
\details{
Similarity is evaluated at the level of min_diff_peak2peak, given in seconds.
The output is an interger vector containing indices of rows. Indices refer to a peak (i.e. substance)
that is similar to the previous peak For example index 5 indicates that the pair 4& 5 is similar.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}
\keyword{internal}

