% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rows.R
\name{similar_rows}
\alias{similar_rows}
\title{Determine which currently distinct substances have similar retention times}
\usage{
similar_rows(average_rts, min_distance = 0.05)
}
\arguments{
\item{average_rts}{\code{vector} of average retention times per row across individuals}

\item{min_distance}{\code{numeric} value defining up to which difference in retention time,
substances are classified as similar}
}
\value{
\item{similar}{\code{vector} of indices pointing to the last index of a row-pair that is similar}
}
\description{
Estimate degree of similarity between subsequent rows by comparison of mean retention times of rows.
}
\details{
Similarity is evaluated at the level of min_distance, given in seconds.
         The output is a interger vector containing indices of rows. Indices refer to a row (i.e. substance)
         that is similar to the previous row. For example index 5 indicates that the pair 4& 5 is similar.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

