% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_chroma.R
\name{simple_chroma}
\alias{simple_chroma}
\title{Simulate chromatograms}
\usage{
simple_chroma(peaks = c(10, 13, 25, 37, 50), N = 1, min = 0, max = 30,
  Names = NULL, sd = NULL)
}
\arguments{
\item{peaks}{vector of mean retention times representing peaks in a chromatogram}

\item{N}{a integer giving the number of chromatograms}

\item{min}{a numeric giving the minimum retention time}

\item{max}{a numeric giving the maximum retention time}

\item{Names}{a character vector giving sample names. By default names are generated by the function}

\item{sd}{a numeric vector of the same length as peaks giving the standard deviation of peaks. Only supported if N = 1. Default is NULL}
}
\value{
a data frame containing x and y coordinates and corresponding sample names
}
\description{
Creates chromatograms with user defined peaks for illustrative purposes
}
\examples{
## create a chromatogram
x <- simple_chroma(peaks = c(5,10,15), N = 1, min = 0, max = 30, Names = "MyChroma")
## plot chromatogram
with(x, plot(x,y, xlab = "time", ylab = "intensity"))

}
