% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.GCalign.R
\name{summary.GCalign}
\alias{summary.GCalign}
\title{Summarizing Peak Alignments with GCalignR}
\usage{
\method{summary}{GCalign}(object, write_txt_file = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{GCalign}, a result of a call to \code{\link{align_chromatograms}}}

\item{write_txt_file}{logical. If \code{TRUE} the output is passed to \code{.txt} file. No text is printed
to the console in this case.}

\item{...}{optional arguments affecting the summary produced.}
}
\description{
summary method for class "GCalign"
}
\examples{
summary(gc_peaks_aligned) # prints summary to the Console

}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
 (meinolf.ottensmann@web.de)
}

