% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_chroma.R
\name{var_per_row}
\alias{var_per_row}
\title{calculates mean retention time per row}
\usage{
var_per_row(chromatograms, rt_col_name)
}
\arguments{
\item{chromatograms}{\code{data.frame} containing GC data (retention times, peak area, peak hight etc) for
one individual in adjacent columns. The first column for all individuals has to be the retention
time, retention time has to be named RT.}
}
\value{
vector with variance of retention times per row in chromatograms
}
\description{
calculates mean retention time per row
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

