% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_transformation.R
\name{adj_ret_time}
\alias{adj_ret_time}
\title{Apply the estimated shift of the retention time to the selected chromatogram
to maximize the similarity compared to the reference}
\usage{
adj_ret_time(Data, OptimalShift, ret_col_name)
}
\arguments{
\item{list, }{output from shared_peaks}
}
\value{
list of number of peaks and shift step for which this is maximised
}
\description{
Apply the estimated shift of the retention time to the selected chromatogram
to maximize the similarity compared to the reference
}
\details{
If competing shifts exists (i.e. same number of peaks are shared),
         selects the smallest absolute value of a shift
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
  (meinolf.ottensmann@web.de)
}

