% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_chromatograms.R
\name{align_chromatograms}
\alias{align_chromatograms}
\title{Aligning chromatograms based on retention times}
\usage{
align_chromatograms(datafile, sep = "\\t", rt_name = NULL,
  write_output = NULL, rt_cutoff_low = NULL, rt_cutoff_high = NULL,
  reference = NULL, step1_maxshift = 0.05, step2_maxshift = 0.02,
  step3_maxdiff = 0.05, blanks = NULL, del_single_sub = FALSE)
}
\arguments{
\item{datafile}{datafile is a tab-delimited txt file. The first rows needs to contain
sample names, the second row column names of the corresponding chromatograms. Starting with
sample IDs, the second row column names of the corresponding chromatograms. Starting with
the third row chromatograms are included, whereby single samples are concatenated horizontally
Each chromatogram needs to consist of the same number of columns, at least
two are required (the retention time and the area)}

\item{rt_name}{character, name of the column holding retention times}

\item{rt_cutoff_low}{numeric, lower threshold under which retention times are cutted}

\item{rt_cutoff_high}{numeric, upper threshold above which retention times are cutted}

\item{reference}{character, a sample to which all other samples are aligned to by means of a
linear shift}

\item{step1_maxshift}{numeric, defines a window to search for an optimal linear shift of samples
with respect to the reference. Shifts are evaluated within - step1_maxshift:step1_maxshift}

\item{step2_maxshift}{numeric, defines the allowed deviation of retention times around the mean
<<<<<<< HEAD
of the corresponding  row}
=======
of the corresponding row}
>>>>>>> 3579af0443af1e4f7aec0512dbbf7a8eaf9ca15f

\item{step3_maxdiff}{numeric, defines the minimum difference in retention times among distict
substances. Substances that do not differ enough, are merged if applicable}

\item{blanks}{character vector of blanks. If specified, all substance found in any of the blanks
will be removed from all samples}

\item{write.output}{character. If specified the output is written to a text file}

\item{delete_single_sub}{logical, determines whether substances that occur in just one sample
are removed or not}
}
\value{

}
\description{
Aligning chromatograms based on retention times
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}
\references{

}

