% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_chromatograms.R
\name{align_chromatograms}
\alias{align_chromatograms}
\title{Aligning chromatograms based on retention times}
\usage{
align_chromatograms(datafile, rt_name = NULL, write_output = NULL,
  rt_cutoff_low = NULL, rt_cutoff_high = NULL, reference = NULL,
  step1_maxshift = 0.05, step2_maxshift = 0.02, step3_maxdiff = 0.05,
  blanks = NULL, del_single_sub = FALSE)
}
\arguments{
\item{datafile}{datafile is a tab-delimited txt file. The first rows needs to contain
<<<<<<< HEAD
sample ID´s, the second row column names of the corresponding chromatograms. Starting with
=======
sample IDs, the second row column names of the corresponding chromatograms. Starting with
>>>>>>> 2605e41ff901173b66a8ad9cb24c30de43a5a015
the third row chromatograms are included, whereby single samples are concatenated horizontally
Each chromatogram needs to consist of the same number of columns, at least
two are required (the retention time and the area)}

\item{rt_name}{character, name of the column holding retention times}

\item{rt_cutoff_low}{numeric, lower threshold under which retention times are cutted}

\item{rt_cutoff_high}{numeric, upper threshold above which retention times are cutted}

\item{reference}{character, a sample to which all other samples are aligned to by means of a
linear shift}
}
\value{

}
\description{
Aligning chromatograms based on retention times
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}
\references{

}

