% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_chromatograms.R
\name{align_chromatograms}
\alias{align_chromatograms}
\title{Aligning chromatograms based on retention times}
\usage{
align_chromatograms(datafile, sep = "\\t", rt_name = NULL,
  write_output = NULL, rt_cutoff_low = NULL, rt_cutoff_high = NULL,
  reference = NULL, step1_maxshift = 0.05, step2_maxshift = 0.02,
  step3_maxdiff = 0.05, blanks = NULL, del_single_sub = FALSE)
}
\arguments{
\item{datafile}{path to the datafile. It has to be a .txt file. The first row needs to contain
sample names, the second row column names of the corresponding chromatograms. Starting with the
third row chromatograms are included, whereby single samples are concatenated horizontally Each
chromatogram needs to consist of the same number of columns, at least two are required (the
retention time and the area).}

\item{sep}{the field separator character. Values on each line of the file are separated by this
character. The default is tab seperated (sep = "\t"). See sep argument in read.table for details.}

\item{rt_name}{Name of the column holding retention times (i.e. "RT")}

\item{write_output}{If specified the aligned output is written to a text file. (i.e. write_output = c("RT", "area")
to write the aligned matrices of the retention times and areas to txt files (the names have
to correspond to the column names given in the second row of the datafile).}

\item{rt_cutoff_low}{lower threshold under which retention times are cutted (i.e. 5)}

\item{rt_cutoff_high}{Upper threshold above which retention times are cutted (i.e. 35)}

\item{reference}{A sample(name) to which all other samples are aligned to by means of a
linear shift (i.e. "individual3") The name has to correspond to an individual name given
in the first line of the datafile.}

\item{step1_maxshift}{Defines a window to search for an optimal linear shift of samples
with respect to the reference. Shifts are evaluated within - step1_maxshift to + step1_maxshift.
Default is 0.05.}

\item{step2_maxshift}{Defines the allowed deviation of retention times around the mean of
the corresponding row. Default is 0.02.}

\item{step3_maxdiff}{Defines the minimum difference in retention times among distinct
substances. Substances that do not differ enough, are merged if applicable. Defailt is 0.05.}

\item{blanks}{character vector of blanks. If specified, all substance found in any of the blanks
will be removed from all samples}

\item{blanks}{Character vector of the names of blanks. If specified, all substance found in any of the blanks
will be removed from all samples (i.e. c("blank1", "blank2")). The names have to correspond
to a name given in the first line of the datafile.}

\item{delete_single_sub}{logical, determines whether substances that occur in just one sample are
removed or not.}
}
\value{

}
\description{
Aligning chromatograms based on retention times
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
 (meinolf.ottensmann@web.de)
}
\references{

}

