% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_chromatograms.R
\name{align_chromatograms}
\alias{align_chromatograms}
\title{Aligning chromatograms based on retention times}
\usage{
align_chromatograms(datafile, sep = "\\t", conc_col_name = NULL,
  rt_col_name = NULL, write_output = NULL, rt_cutoff_low = NULL,
  rt_cutoff_high = NULL, reference = NULL, max_linear_shift = 0.05,
  max_diff_peak2mean = 0.02, min_diff_peak2peak = 0.03, blanks = NULL,
  delete_single_peak = FALSE, n_iter = 1)
}
\arguments{
\item{datafile}{path to a datafile. It has to be a .txt file. The first row needs to contain
sample names, the second row column names of the corresponding chromatograms. Starting with the
third row chromatograms are included, whereby single samples are concatenated horizontally. Each
chromatogram needs to consist of the same number of columns, at least two are required:The
retention time and a measure of concentration (e.g. peak area or height).}

\item{sep}{the field separator character. Values on each line of the file are separated by this
character. The default is tab seperated (sep = '\\t'). See sep argument in read.table for details.}

\item{conc_col_name}{character string naming a column used for quantification of peaks (e.g. peak area or peak height).
The designated variable needs to be numeric.}

\item{rt_col_name}{character string naming the column holding retention times.The designated variable needs to be numeric.}

\item{write_output}{character vector of variables to write to a text file (e.g. \code{c("RT","Area")}.
The default is \code{NULL}.Names of the text files are concatenations of \code{datafile} and the output variables.}

\item{rt_cutoff_low}{lower threshold under which retention times are cutted (i.e. 5)}

\item{rt_cutoff_high}{upper threshold above which retention times are cutted (i.e. 35)}

\item{reference}{\code{character}, a sample(name) to which all other samples are aligned to by means of a
linear shift (i.e. "individual3") The name has to correspond to an individual name given
in the first line of \code{datafile}.}

\item{max_linear_shift}{defines a window to search for an optimal linear shift of samples
with respect to the reference. Shifts are evaluated within - max_linear_shift to + max_linear_shift.
Default is 0.05.}

\item{max_diff_peak2mean}{defines the allowed deviation of retention times around the mean of
the corresponding row. Default is 0.02.}

\item{min_diff_peak2peak}{defines the minimum difference in retention times among distinct
substances. Substances that differ less, are merged if every sample contains either one
or none of the respective compounds. This parameter is a major determinant in the classification
of distinct peaks. Therefore careful consideration is required to adjust this setting to
your needs (i.e. the resolution of your gas-chromatography pipeline). Default is 0.02.}

\item{blanks}{character vector of names of blanks. If specified, all substances found in any of the blanks
will be removed from all samples (i.e. c("blank1", "blank2")). The names have to correspond
to a name given in the first line of the datafile.}

\item{delete_single_peak}{logical, determines whether substances that occur in just one sample are
removed or not. By default single substances are retained in chromatograms.}

\item{n_iter}{integer indicating the iterations of the core alignment algorithm.}
}
\value{
Returns an object of class GCalign that is a a list with the following elements:
\item{call}{function call}
\item{chroma_aligned}{a list containing data.frames with the aligned variables}.
\item{rt_raw}{a data.frame with the retention times before alignment}
\item{rt_linear}{a data.frame with the retention times after the linear transformation}
\item{rt_aligned}{a data.frame with the final aligned retention times}
\item{parameter}{a data.frame}{containing the arguments of the function call}
}
\description{
\code{align_chromatograms()} is the core function of \link{GcalignR}.
}
\details{
Alignment is archieved by running three major algorithms always considering the complete
set of samples submitted to the function. In brief: All chromatograms are linearly shifted to
maximise similarity with a common reference to account for systematic shifts in retention times
caused by the gas-chromatography setup. Second peaks of similar retention times are step-wise
moved to the same location (i.e row) to group substances. In a third step peaks (i.e.) substances
that show smaller differences in mean retention times than expected by the achievable resolution
of gas-chromatography or the chemistry of the compounds are merged. Optional processing includes the
removal of peaks present in blanks (i.e. contaminations) and peaks that a uniquely found within just
one sample.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
 (meinolf.ottensmann@web.de)
}

