% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_chromatograms.R
\name{align_chromatograms}
\alias{align_chromatograms}
\title{Aligning Peaks based on retention times}
\usage{
align_chromatograms(data, sep = "\\t", conc_col_name = NULL,
  rt_col_name = NULL, write_output = NULL, rt_cutoff_low = NULL,
  rt_cutoff_high = NULL, reference = NULL, max_linear_shift = 0.05,
  max_diff_peak2mean = 0.02, min_diff_peak2peak = 0.02, blanks = NULL,
  delete_single_peak = FALSE, n_iter = 1, merge_rare_peaks = FALSE)
}
\arguments{
\item{data}{Two options. (1) The most common input is the path to a file with extension \code{.txt} to load data.
The first row needs to contain sample names, the second row column names of the corresponding chromatograms. Starting with the
third row, peak data are included, whereby matrices of single samples are concatenated horizontally. The
matrix for each sample needs to consist of the same number of columns, at least two are required:The
retention time and a measure of concentration (e.g. peak area or height). See the package
vignette for an example. (2) The alternative input is a list of data.frames. Each data.frame
contains the peak data for a single individual with at least two variables, the retention time of
the peak and the area under the peak. The variables have to have the same name across all samples
(data.frames). Also, each list element (i.e. each data.frame) has to be named with the ID of
the individual. See the vignette for an example.}

\item{sep}{The field separator character. Values on each line of the file are separated by this
character. The default is tab seperated (sep = '\\t'). See the \code{sep} argument in \code{\link[utils]{read.table}} for details.}

\item{conc_col_name}{Character string naming a column used for the quantification of peaks (e.g. peak area or peak height).
The designated variable needs to be numeric.}

\item{rt_col_name}{Character string naming the column containing retention times.The designated variable needs to be numeric.}

\item{write_output}{Character vector of variables to write to a text file (e.g. \code{c("RT","Area")}.
The default is \code{NULL}.Names of the text files are concatenations of \code{datafile} and the output variables.
Strings need to correspond to column names of the \code{datafile}.}

\item{rt_cutoff_low}{Lower threshold under which retention times are cutted (i.e. 5 minutes). Default is NULL.}

\item{rt_cutoff_high}{Upper threshold above which retention times are cutted (i.e. 35 minutes). Default is NULL.}

\item{reference}{Character string of a sample to which all other samples are aligned to by means of a
linear shift (e.g. \code{"individual3"}. The name has to correspond to an individual name given
in the first line of \code{data}. Alternatively a sample called \code{reference} can be included
in \code{data} containing user-defined peaks (e.g. an internal standard) to align the samples to. After the linear
transformation \code{reference} will be removed from the data.}

\item{max_linear_shift}{Defines a window to search for an optimal linear shift of chromatogram peaks
with respect to the reference. Shifts are evaluated within - \code{max_linear_shift to + max_linear_shift}.
Default is 0.05.}

\item{max_diff_peak2mean}{Defines the allowed deviation of retention times around the mean of
the corresponding row. Peaks with differing retention times are moved
to a more appropriate mean retention time. Default is 0.02.}

\item{min_diff_peak2peak}{Defines the minimum difference in retention times among distinct
substances. Substances that differ less, are merged if every sample contains either one
or none of the respective compounds. This parameter is a major determinant in the classification
of distinct peaks. Therefore careful consideration is required to adjust this setting to
your needs (i.e. the resolution of your gas-chromatography pipeline). Large values might cause the
merge of true peaks, if those are not occuring within one individual, which might happen by chance
with a low sample size. Small values can be considered as conservative. Default is 0.02.}

\item{blanks}{Character vector of names of blanks. If specified, all substances found in any of the blanks
will be removed from all samples (i.e. c("blank1", "blank2")). The names have to correspond
to a name given in the first line of \code{data}.}

\item{delete_single_peak}{logical, determines whether substances that occur in just one sample are
removed or not. By default single substances are retained in chromatograms.}

\item{n_iter}{integer indicating the number of iterations of the core alignment algorithm. Additional replications of
the alignment and merging steps can be helpful to clean-up chromatograms, that otherwise show
some remaining peak outliers mapped to the wrong mean retention time. Inspect alignment visually
with a Heatmap \code{\link{gc_heatmap}}.}

\item{merge_rare_peaks}{logical determining whether peaks that are redundant by means of a similar retention times
are merged. If \code{TRUE} peaks are merged as long as only 5 % of the samples contain two peaks.
Always the peak with the higher abundance (i.e. peak area or peak height) is retained.}
}
\value{
Returns an object of class GCalign that is a a list with the following elements:
\item{call}{function call}
\item{chroma_aligned}{a list containing data.frames with the aligned variables}
\item{rt_raw}{a data.frame with the retention times before alignment}
\item{rt_linear}{a data.frame with the retention times after the linear transformation}
\item{rt_aligned}{a data.frame with the final aligned retention times}
\item{align_summary}{a data.frame}{A basic summary of the alignment process}

 @author Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
 (meinolf.ottensmann@web.de)
}
\description{
\code{align_chromatograms()} is the core function of \code{\link{GCalignR}}.
}
\details{
The Alignment of Peaks is archieved by running three major algorithms always considering the complete
set of samples submitted to the function. In brief: All peaks are linearly shifted to
maximise similarity with a reference to account for systematic shifts in retention times
caused by gas-chromatography processing. Second, peaks of similar retention times are
transfered step-wise to the same location (i.e row) in order to group substances. In a third step peaks (i.e.substances)
that show smaller differences in mean retention times than expected by the achievable resolution
of the gas-chromatography or the chemistry of the compounds are merged. Several optional processing steps are available
ranging from the removal of peaks representing contaminations (requires to include blanks as a control) to the removal
of uninformative peaks that are present in just one sample.
}
\examples{
data(gc_peak_data)
gc_peak_data <- gc_peak_data[1:4]
out <- align_chromatograms(gc_peak_data, conc_col_name = "area", rt_col_name = "RT",
       rt_cutoff_low = 5, rt_cutoff_high = 45, reference = "ind3",
         max_linear_shift = 0.05, max_diff_peak2mean = 0.02, min_diff_peak2peak = 0.03,
         blanks = NULL, delete_single_peak = TRUE)

}

