% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_var.R
\name{align_var}
\alias{align_var}
\title{Quantify goodness of alignment by relative variances of peak retention times}
\usage{
align_var(gc_peak_list, rt_col_name)
}
\arguments{
\item{gc_peak_list}{List of data.frames. Each data.frame contains GC-data (e.g. retention time, peak area, peak height) of one sample. Variables are stored in columns.
Rows represent distinct peaks. Retention time is a required variable.}

\item{rt_col_name}{character string naming the column holding retention times.The designated variable needs to be numeric.}
}
\value{
a list of two numeric vectors
\item{range}{the minimum and maximum relative variation of peak retention times}
\item{average}{the mean relative variation of peak retention times}
}
\description{
\code{align_var()} calculates range and mean of the coefficient of variation for retention
times of peaks that are currently within the same row in data.frames of \code{gc_peak_list}
(i.e. classified as same substance).
}

