% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_transformation.R
\name{best_shift}
\alias{best_shift}
\title{Find the best linear shift}
\usage{
best_shift(peak_list)
}
\arguments{
\item{peak_list}{data.frame, the output from shared_peaks}
}
\value{
optimal shift in seconds to yield the maximum number of shared peaks
}
\description{
Selects the optimal shifting time leading to the maximum number of shared peaks
}
\details{
If competing shifts exists (i.e. same number of peaks are shared),
         selects the smallest absolute value of a shift
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
  (meinolf.ottensmann@web.de)
}

