% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_shift.R
\name{best_shift}
\alias{best_shift}
\title{Find the best linear shift}
\usage{
best_shift(peaks)
}
\arguments{
\item{peaks}{data.frame including a shift value and the corresponding number of shared peaks.
Generated by \code{\link{shared_peaks}}}
}
\value{
optimal shift in seconds to yield the maximum number of shared peaks
}
\description{
Selects the optimal shifting time leading to the maximum number of shared peaks
between a chromatogram and its reference.
}
\details{
If competing shifts exists (i.e. same number of peaks are shared),
selects the smallest absolute value of a shift
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
  (meinolf.ottensmann@web.de)
}

