% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{Check correct input format for processing in GCalignR}
\usage{
check_input(data, show_peaks = FALSE, sep = "\\t", ...)
}
\arguments{
\item{data}{path to a data file or the name of a list in the Global Environment.}

\item{show_peaks}{logical, if TRUE the distribution of peak numbers is plotted. Default is FALSE.}

\item{sep}{The field separator character. The default is tab seperated (\code{sep = '\\t'}). See the "sep" argument in \code{\link[utils]{read.table}} for details.}

\item{...}{optional arguments passed to methods, see \code{\link[graphics]{barplot}}. Only used if
\code{show_peaks==TRUE}.}
}
\description{
Checks conformity between the input format and the requirements of GCalignR. Supported are
a path to a Text file (i.e. "data.txt") or a list of data frames. See \code{\link{align_chromatograms}} for details.
}
\examples{
check_input(peak_data) ## Checks format
check_input(peak_data, show_peaks=TRUE) ## Includes a barplot of peaks

}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
 (meinolf.ottensmann@web.de)
}

