% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{Check correct formatting of data input}
\usage{
check_input(data, plot_peak_distribution = FALSE, sep = "\\t", ...)
}
\arguments{
\item{data}{path to a data file or the name of a list in the Global Environment.}

\item{plot_peak_distribution}{logical, if TRUE the distribution of peak numbers is plotted. Default is FALSE}

\item{sep}{The field separator character. Values on each line of the file are separated by this
character. The default is tab seperated (sep = '\\t'). See \code{sep} argument in \code{\link[utils]{read.table}} for details.}

\item{...}{optional arguments passed to methods, see \code{\link[graphics]{barplot}}. Only used if
\code{plot_peak_distribution==TRUE}.}
}
\value{
TRUE if data is formatted correctly, warning and explanation if not.
}
\description{
Checks conformity between the input format and the requirements of GCalignR. Supported are
a \code{.txt} files or lists of data.frames. See \code{\link{align_chromatograms}} for details.
}
\examples{
data(gc_peak_data)
gc_peak_data <- gc_peak_data[1:4]
check_input(gc_peak_data)

}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
 (meinolf.ottensmann@web.de)
}

