% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_redundancy.R
\name{check_redundancy}
\alias{check_redundancy}
\title{Check whether similar rows are redundant or not}
\usage{
check_redundancy(gc_peak_df, similar_peaks, rt_col_name)
}
\arguments{
\item{gc_peak_df}{data.frame containing GC-data (e.g. retention time, peak area, peak height) of one sample. Variables are stored in columns.
Rows represent distinct peaks. Retention time is a required variable.}

\item{similar_peaks}{integer indexing the position of a peaks that is similar to the one before. Output from \code{similar_peaks}.}

\item{rt_col_name}{character string naming the column holding retention times.The designated variable needs to be numeric.}
}
\value{
a binary vector of ones and zeros. One indicates redundancy.
}
\description{
\code{check_redundancy()} evaluates if peaks of similar retention time are redundant.
This is ture as long as any sample contains peaks in either one or none of the
row-pair, but never in both.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

