% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rows.R
\name{check_redundancy}
\alias{check_redundancy}
\title{If only one of two neighbouring rows contain a substancethey are redundant, coded by a One}
\usage{
check_redundancy(chromatogram, similar)
}
\arguments{
\item{similar}{}
}
\value{
if similar rows are redundant, 1, else 0
}
\description{
If only one of two neighbouring rows contain a substancethey are redundant, coded by a One
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

