% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_gc_mat_to_list.R
\name{conv_gc_mat_to_list}
\alias{conv_gc_mat_to_list}
\title{Transforms GC-data to list}
\usage{
conv_gc_mat_to_list(gc_data, ind_names, var_names)
}
\arguments{
\item{gc_data}{data.frame containing GC-data for all individuals in adjacent columns. Samples need to have the same
number of columns (i.e variables) in the same the order.}

\item{ind_names}{character vector with names of individuals following the same order as in \code{gc_data}.}

\item{var_names}{character vector with names of columns from gc_data. For example:
\code{var_names =c("RT", "Area", "Height")}.}
}
\value{
A list of data.frames containing GC-data of individual samples
}
\description{
Transforms a data.frame containing GC-data (e.g. retention times, peak area, peak height etc.) to a list.
Data belonging to individual samples are concatenated horizontally in \code{gc_data}. Every samples needs to
have the same number of columns in consistent order. This means for example columns 1:3 belong to \code{Ind A}
and columns 4.6 to \code{Ind B} and so forth.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

