% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_empty_rows.R
\name{delete_empty_rows}
\alias{delete_empty_rows}
\title{Delete empty rows}
\usage{
delete_empty_rows(gc_peak_df, average_rts)
}
\arguments{
\item{gc_peak_df}{data.frame containing GC-data (e.g. retention time, peak area, peak height) of one sample. Variables are stored in columns.
Rows represent distinct peaks. Retention time is a required variable.}

\item{average_rts}{numeric}
}
\description{
\code{del_empty_rows} deletes all rows of data.frame that are empty. This is true for rows
with an average retention time of value NA (i.e. no peak is present in any data.frame) at this
location.
}

