% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_cols.R
\name{delete_space_colnames}
\alias{delete_space_colnames}
\title{Deletes spaces when renaming table columns in GC list}
\usage{
delete_space_colnames(data)
}
\arguments{
\item{data}{\code{data.frame} containing GC data (retention times, peak area, peak hight etc) for
one individual in adjacent columns. The first column for all individuals has to be the retention
time}
}
\value{
List of data.frames. Each list element is the GC data for one individual
}
\description{
Deletes spaces when renaming table columns in GC list
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}
\references{

}

