% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_space_colnames.R
\name{delete_space_colnames}
\alias{delete_space_colnames}
\title{Remove spaces in column names}
\usage{
delete_space_colnames(gc_data)
}
\arguments{
\item{gc_data}{data.frame containing GC-data for all individuals in adjacent columns. Samples need to have the same
number of columns (i.e variables) in the same the order.}
}
\value{
data.frame containing GC data with removed spaces in column names
}
\description{
If present, spaces within column names are eliminated
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

