% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC_Heatmap.R
\name{GC_Heatmap}
\alias{GC_Heatmap}
\title{GC_Heatmap visualises the goodness of chromatogram alignment}
\usage{
<<<<<<< HEAD
GC_Heatmap(GcOut, all = FALSE)
}
\arguments{
\item{GcOut}{\code{data.frame} representing a matrix of retention times,
where samples are ordered in rows}

\item{all}{logical, indicating whether all or just the final retention times
are plotted. Default \code{all=FALSE}}
=======
GC_Heatmap(rt_df)
}
\arguments{
\item{rt_df}{Dataframe}
>>>>>>> 89576ca767e3575ff20d45eaf308c23e3efc35f6
}
\value{

}
\description{
Visualises the deviation of retention of sample from the mean of all other,
 to indicate how well chromatograms are aligned. Small deviations indicate a good alignment
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}
\references{

}

