% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC_Heatmap.R
\name{gc_heatmap}
\alias{gc_heatmap}
\title{Visualises peak alignments in form of a heatmap}
\usage{
gc_heatmap(object = NULL, algorithm_step = c("aligned", "linear_shifted",
  "pre_alignment"), substance_subset = NULL, legend_type = c("legend",
  "colourbar"), samples_subset = NULL, type = c("binary", "discrete"),
  threshold = NULL, label_size = NULL, show_legend = TRUE,
  main_title = NULL, label = c("y", "xy", "x", "none"))
}
\arguments{
\item{object}{Object of class "GCalign", the output of a call to \link{align_chromatograms}.}

\item{algorithm_step}{Character indicating which step of the algorithm is plotted. Either \strong{"pre_alignment"}, \strong{"linear_shifted"} or \strong{"aligned"} specifying the raw, linearly shifted or aligned data respectively. Default is the heatmap for the aligned data.}

\item{substance_subset}{A vector of integers containing indices of substances in ascending order of retention times to plot. By default all substances are plotted.}

\item{legend_type}{A character specifying the type of colourbar as \strong{discrete} (i.e retention times are classified as deviating or not) or \strong{gradient} to present deviations on a fine scale).}

\item{samples_subset}{A vector indicating which samples are plotted on the heatmap.
Either a numeric vector of indices (order in the input) or a vector of sample names.}

\item{type}{A character specifying whether a \strong{'binary'} heatmap or a heatmap of \strong{'discrete'}
deviations is plotted.}

\item{threshold}{A numeric value denoting the threshold above which the deviation of individual peak retention times
from the mean retention time of the respective substance is highlighted in \emph{binary} heatmaps. By default, the value of parameter \code{max_diff_peak2mean} (see \code{\link{align_chromatograms}}) that was used in aligning the data is used.}

\item{label_size}{An integer determining the size of labels on y and x axis. By default a fitting label_size is calculate (beta!) to compromise between readability and messiness due to a potentially large number of substances and samples. Note: Labels for substances on the x axis are only possible if a maximum of 150 substances are plotted.}

\item{show_legend}{Logical determining whether a legend is included or not. Default is TRUE.}

\item{main_title}{Character giving the title of the heatmap. If not specified, titles are generated automatically.}

\item{label}{character determining if labels are shown on axis. Depending on the number of peaks and/or samples, labels are difficult to read. Use subsets instead. Possible option are \strong{"xy", "x", "y" or "none"}}
}
\value{
object of class "ggplot"
}
\description{
The goal of aligning peaks is to group homologous peaks (ideally representing homologous substances) in the same row across samples, although peaks have slightly different retention times across samples. This function makes it possible to evaluate the alignment quickly by inspecting the (i) distribution of peaks across samples, (ii) the variation within rows as well as (iii) patterns that might hint at splitting peaks across rows. The mean retention time per row is here defined as the "true"  retention time and deviations of individual peaks can be seen by a large deviation in the retention time to the mean. Subsetting of the retention time range (i.e. selecting peaks by the mean retention time) and samples (by name or by position) allow to quickly inspect region of interest. Two types of heatmaps are available, a binary heatmap allows to determine if the retention time of single samples deviates by more than the user defined threshold from the mean. Optionally, a discrete heatmap allows to check deviations quantitatively. Large deviation can have mutliple reasons. The most likely explaination is given by the fact that adjacent rows were merged as specified by the value \code{min_diff_peak2peak} in \code{\link{align_chromatograms}}. Here clear cases, in which peaks of multiple samples have been grouped in either of two or more rows can be merged and cause relatively high variation in peak retention times.
}
\examples{

 ## aligned gc-dataset
 data("aligned_peak_data")
 ## Default settings: The final output is plotted
 gc_heatmap(aligned_peak_data, algorithm_step = "aligned")

 ## Plot the input data
 gc_heatmap(aligned_peak_data,algorithm_step = "pre_alignment")

 ## Plot a subset of the first 50 scored substances
 gc_heatmap(aligned_peak_data,algorithm_step="aligned",substance_subset = 1:50)

 ## Plot specific samples, apply a stricter threshold
 gc_heatmap(aligned_peak_data,samples_subset = c("M2","P7","M13","P13"),threshold = 0.02)

}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}
