% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC_Heatmap.R
\name{GC_Heatmap}
\alias{GC_Heatmap}
\title{Visualise deviation of retention times from the mean per row
gc_heatmap visualises the current variation of retention times among all samples
within the same row. Therefore the absolute deviation from the mean of the row is calculated
for each sample}
\usage{
GC_Heatmap(rt_df)
}
\arguments{
\item{rt_df}{Dataframe}
}
\value{

}
\description{
Visualise deviation of retention times from the mean per row
gc_heatmap visualises the current variation of retention times among all samples
within the same row. Therefore the absolute deviation from the mean of the row is calculated
for each sample
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}
\references{

}

