% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC_Heatmap.R
\name{GC_Heatmap}
\alias{GC_Heatmap}
\title{GC_Heatmap visualises the goodness of chromatogram alignment}
\usage{
GC_Heatmap(GcOut, step = "rt_aligned", substance_subset = NULL,
  guide = "colourbar", limits = c(-0.05, 0.05))
}
\arguments{
\item{GcOut}{\code{data.frame} representing a matrix of retention times,
where samples are ordered in rows}

\item{step}{character, indicating which step of the algorithm to plot. Either
{rt_raw}, {rt_linear} or {rt_aligned}}

\item{substance_subset}{vector containing indices of substances (i.e. rows) to plot
By default {NULL} indicating all substances are plotted}

\item{guide}{character, indicating type of colourbar as discrete (i.e 'legend')
or gradient (i.e 'colourbar)}

\item{limits}{vector, allows to finetune the scaling of the colourbar}
}
\value{

}
\description{
Visualises the deviation of retention of sample from the mean of all other,
 to indicate how well chromatograms are aligned. Small deviations indicate a good alignment
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}
\references{

}

