% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC_Heatmap.R
\name{gc_heatmap}
\alias{gc_heatmap}
\title{Visualises peak alignments in form of a heatmap}
\usage{
gc_heatmap(object, algorithm_step = c("aligned", "linear_shifted",
  "pre_alignment"), substance_subset = NULL, legend_type = c("legend",
  "colourbar"), samples_subset = NULL, type = c("binary", "discrete"),
  threshold = 0.05, label_size = NULL, show_legend = TRUE,
  main_title = NULL, label = TRUE)
}
\arguments{
\item{object}{Object of class "GCalign", the output of a call to \link{align_chromatograms}.}

\item{algorithm_step}{Character indicating which step of the algorithm is plotted. Either \strong{"pre_alignment"}, \strong{"linear_shifted"} or \strong{"aligned"} specifying the raw, linearly shifted or aligned data respectively. Default is the heatmap for the aligned data.}

\item{substance_subset}{A vector of integers containing indices of substances in ascending order of retention times to plot. By default all substances are plotted.}

\item{legend_type}{A character specifying the type of colourbar as \strong{discrete} (i.e retention times are classified as deviating or not) or \strong{gradient} to present deviations on a fine scale).}

\item{samples_subset}{A vector indicating which samples are plotted on the heatmap.
Either a numeric vector of indices (order in the input) or a vector of sample names.}

\item{type}{A character specifying whether a \strong{'binary'} heatmap or a heatmap of \strong{'discrete'}
deviations is plotted.}

\item{threshold}{A numeric value denoting the threshold above which the deviation of individual peak retention times
from the mean retention time of the respective substance is highlighted in \emph{binary} heatmaps.}

\item{label_size}{An integer determining the size of labels on y and x axis. By default a fitting label_size is calculate (beta!) to compromise between readability and messiness due to a potentially large number of substances and samples. Note: Labels for substances on the x axis are only possible if a maximum of 150 substances are plotted.}

\item{show_legend}{Logical determining whether a legend is included or not. Default is TRUE.}

\item{main_title}{Character giving the title of the heatmap. If not specified, titles are generated automatically.}

\item{label}{Logical determining whether labels for samples and substances are used. By default sample names are written on the vertical axis, but in large data sets its less readable and may be used effectively only for subsets of the data.}
}
\value{
object of class "ggplot"
}
\description{
The goal of aligning chromatography peaks is to get the same substance in the same row, although it
might have slightly different retention times across samples. This function makes it possible
to evaluate the alignment by illustrating how far the retention time of a given peak within a sample
deviates from the mean retention time for this substance across all samples after alignment. In other words,
given that the mean retention time is the 'real' retention time of a substance, the heatmap
shows how far a given peak deviates from it. Moreover, subsetting both, retention time range and samples allow to quickly inspect the distribution of peaks in order to evaluate the alignment. Two types of heatmaps are available. A binary heatmap allows to determine
if the retention time of single samples deviates by more than a user defined threshold from the mean. Optionally, a discrete heatmap allows to check deviations quantitatively.
}
\examples{

 ## aligned gc-dataset
 data("aligned_peak_data")
 ## Default settings: The final output is plotted
 gc_heatmap(aligned_peak_data, algorithm_step = "aligned")

 ## Plot the input data
 gc_heatmap(aligned_peak_data,algorithm_step = "pre_alignment")

 ## Plot a subset of the first 50 scored substances
 gc_heatmap(aligned_peak_data,algorithm_step="aligned",substance_subset = 1:50)

 ## Plot specific samples, apply a stricter threshold
 gc_heatmap(aligned_peak_data,samples_subset = c("M2","P7","M13","P13"),threshold = 0.02)

}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}
