% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC_Heatmap.R
\name{gc_heatmap}
\alias{gc_heatmap}
\title{gc_heatmap visualises the goodness of the chromatogram alignment}
\usage{
gc_heatmap(GcOut, algorithm_step = c("aligned_rt", "linear_shifted_rt",
  "initial_rt"), substance_subset = NULL, guide = c("legend", "colourbar"),
  samples_subset = NULL, type = c("binary", "continous"),
  threshold = 0.05)
}
\arguments{
\item{GcOut}{\code{data.frame} output of align_chromatograms. Class \code{GCalign}}

\item{algorithm_step}{\code{character} indicating which step of the algorithm is plotted. Either
\code{rt_raw}, \code{rt_linear} or \code{rt_aligned}}

\item{substance_subset}{\code{vector} containing indices of substances (i.e. rows) to plot
By default \code{NULL} indicating all substances are plotted}

\item{guide}{\code{character} indicating type of colourbar as discrete (i.e 'legend')
or gradient (i.e 'colourbar)}

\item{samples_subset}{\code{vector} indicating which samples are plotted on the heatmap.
Either a \code{numeric} \code{vector} of indices or a \code{character} \code{vector}
of sample names}

\item{type}{\code{character} specifying whether a binary heatmap or a heatmap of continous
deviations is plotted.}

\item{threshold}{\code{numeric} indicates the acceptable deviation of individual peak retention times
from the mean retention time of the respective peak across all samples.}
}
\value{
\item{hm}{object of \code{class} ggplot}
}
\description{
Visualises the deviation of single samples from the whole population by comparing
         the actual retention time against the mean of all other samples containing the same
         substance. Two types of heatmaps are available. A binary heatmap allows to determine
         if single samples within each chromatogram are correctly assigned to a certain substance,
         by setting a fixed threshold of allowed deviations from the mean. The optional discrete
         heatmap allows to check the deviations quantitatively.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

