% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC_Heatmap.R
\name{GC_Heatmap}
\alias{GC_Heatmap}
\title{Visualise deviation of retention times from the mean per row
gc_heatmap visualises the current variation of retention times among all samples
within the same row. Therefore the absolute deviation from the mean of the row is calculated
for each sample}
\usage{
GC_Heatmap(rts_matrix)
}
\arguments{
\item{rts_matrix}{Matrix of Retention times, where samples are sorted in rows
and retention times in columns. Absence of substances in a sample are
coded by NA. These values are ignored in calculating the row means}
}
\value{

}
\description{
Visualise deviation of retention times from the mean per row
gc_heatmap visualises the current variation of retention times among all samples
within the same row. Therefore the absolute deviation from the mean of the row is calculated
for each sample
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}
\references{

}

