% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCalignR.R
\docType{package}
\name{GCalignR}
\alias{GCalignR}
\alias{GCalignR-package}
\title{GCalignR: A package to align gas-chromatograms based on retetion times}
\description{
The GCalignR package provides a three step alogrithm to align GC data
}
\section{GCalignR main function}{

}

