% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCalignR.R
\docType{package}
\name{GCalignR}
\alias{GCalignR}
\alias{GCalignR-package}
\title{GCalignR: A package to align gas-chromatograms based on retention times}
\description{
\code{GCalignR} contains the following functions:
\code{\link{align_chromatograms}}
\code{\link{GC_Heatmap}}
}
\details{
gas-chromatography (GC) & gas-chromatography-mass-spectrometry (GC-MS) are
powerful analytical separation techniques that can be used for diverse
applications. Even though gas-chromatography is highly sensitive and retention
times are substance-specific (in equal settings) small systematic shifts in
chromatograms as well as unavoidable inaccuracies in peak detection make it difficult
to compare chromatograms with each other. Researcher often align chromatograms
painful by hand, thereby investing a lot of time in preprocessing their data in
a bias-prone manner.

This package provides a three step alogrithm for a automated alignment of
chromatograms based on objective and reproducible parameters.
....
}

