% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCalignR.R
\docType{package}
\name{GCalignR}
\alias{GCalignR}
\alias{GCalignR-package}
\title{GCalignR: A package to align gas-chromatography peaks based on retention times}
\description{
\strong{\code{GCalignR}} contains the functions listed below. Follow the links to access the documentation of each function.

\code{\link{align_chromatograms}} executes all alignment steps.

\code{\link{check_input}} tests the input data for formatting issues.

\code{\link{draw_chromatogram}} visualises peak lists in form of a chromatogram.

\code{\link{find_peaks}} detects and calculates peak heights in chromatograms. Not intended to be used for peak integration in empirical data. Used for illustration purposes only.

\code{\link{gc_heatmap}} visualises aligned datasets using heatmaps that can be customised.

\code{\link{norm_peaks}} allows to compute the relative abundance of peaks with samples.

\code{\link{peak_interspace}} gives a histogram of the distance between peaks within samples over the whole dataset.

\code{\link{read_peak_list}} reads the content of a text file and converts it to a list.

\code{\link{simple_chroma}} creates simple chromatograms for testing and illustration purposes.
}
\details{
The usage of the package is explained in the vignettes distributed with \strong{GCalignR}. Type \code{browseVignettes("GCalignR")} access them.
}
