% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rows.R
\name{is_redundant}
\alias{is_redundant}
\title{Checks if similar rows (i.e. substances) are redundant}
\usage{
is_redundant(redundant, criterion = "strict")
}
\arguments{
\item{redundant}{\code{vector} of integers coding redundancy as 1=Yes & 0=No for every sample at a given
pair of similar rows.}

\item{criterion}{\code{character} specifying by which method merging will be}
}
\value{
\item{ToMerge}{\code{integer} indicating whether will be merge (1) or not (0)}
}
\description{
Indicates by a binary output variable (1/0) which rows will be merged in the next step
}
\details{
Methods: "strict": A single sample with two peaks prevents merging
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

