% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rows.R
\name{is_redundant}
\alias{is_redundant}
\title{Indicates by a binary output variable (1/0) if rows should be merged}
\usage{
is_redundant(similar, redundant, criterion = "strict")
}
\arguments{
\item{similar}{}
}
\value{
if similar rows are redundant, 1, else 0
}
\description{
Indicates by a binary output variable (1/0) if rows should be merged
}
\details{
Methods: "strict": A single sample with two peaks prevents merging
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

