% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_redundant.R
\name{is_redundant}
\alias{is_redundant}
\title{Checks if similar rows (i.e. substances) are redundant}
\usage{
is_redundant(redundant, criterion = "strict")
}
\arguments{
\item{redundant}{\code{vector} of integers coding redundancy as 1=Yes & 0=No for every sample at a given
pair of similar rows.}

\item{criterion}{character string either "strict" or "proportional". "strict" is used to delete the row containing zeros,
"proportional" provokes the selection of the larger of two peaks, defined by the concentration of the peaks.
\code{criterion="proportional"} will only affect pairs of peaks where less than 5% of the samples containg
two conflicting peaks.}
}
\value{
\item{ToMerge}{\code{integer} indicating whether will be merge (1) or not (0)}
}
\description{
Indicates by a binary output variable (1/0) which rows will be merged in the next step
}
\details{
Methods: "strict": A single sample with two peaks prevents merging
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

