% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_transformation.R
\name{linear_transformation}
\alias{linear_transformation}
\title{Full Alignment of Peak Lists by linear retention time correction.}
\usage{
linear_transformation(gc_peak_list, reference, max_linear_shift = 0.05,
  step_size = 0.01, rt_col_name, Logbook = NULL)
}
\arguments{
\item{gc_peak_list}{List of data.frames. Each data.frame contains GC-data
(e.g. retention time, peak area, peak height) of one sample. Variables are
stored in columns. Rows represent distinct peaks. Retention time is a
required variable.}

\item{reference}{A character giving the name of a sample included in the dataset. All sample are aligned to the reference.}

\item{max_linear_shift}{This value defines the maximum time that one chromatogram is expected to be deviating in retention times
from another chromatogram. To correct for these systematic shifts, the algorithm potentially adds the same value to all peaks within a chromatogram to maximise the number of shared peaks with
the reference. We recommend to start with the default of 0.02 (minutes) and increase if necessary.}

\item{step_size}{Integer giving the step size in which linear shifts are evaluated between \strong{max_linear_shift} and \strong{-max_linear_shift}.}

\item{rt_col_name}{Character string giving the name of the column containing the retention times.The variable needs to
be numeric and the decimal separator needs to be a point.}

\item{Logbook}{A list. If present, a summary of the applied linear shifts in full alignments of peak lists is appended to the list. If not specified, a list will be created automatically.}
}
\value{
A list containing two items.
\item{chroma_aligned}{Transformed data}
\item{Logbook}{Logbook, record of the applied shifts}
List of data.frames containing chromatograms with applied linear shifts
}
\description{
Shifts all peaks within samples to maximise the similarity to a reference sample. For optimal results, a sufficient number of landmark peaks that are shared among samples are required to find the most adequate linear shifts. A reference needs to be selected before, for instance using \code{\link{choose_optimal_reference}}. Linear shifts are evaluated within a user-defined window in discrete steps. The highest similarity score defines the shift that will be applied. If more than a single shift step yield the same similarity score, the smallest absolute value wins in order to avoid overcompensation. The functions is envoked internally by \code{\link{align_chromatograms}}.
}
\details{
A similarity score is calculated as the sum of deviations in retention times between all reference peaks and the closest peak in the sample. The principle idea is that the appropriate linear transformation will reduce the deviation in retention time between homologous peaks, whereas all other peaks should deviate randomly. Among all considered shifts, the minimum deviation score is selected for subsequent full alignment by shifting all peaks of the sample by the same value.
}
\examples{
dat <- peak_data[1:10]
dat <- lapply(dat, function(x) x[1:50,])
x <- linear_transformation(gc_peak_list = dat, reference = "C2", rt_col_name = "time")

}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}
