% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_transformation.R
\name{linear_transformation}
\alias{linear_transformation}
\title{Shift peaks to eliminate systematic inaccuracies of peak detection by GC.}
\usage{
linear_transformation(gc_peak_list, reference, max_linear_shift = 0.05,
  step_size = 0.01, error = 0, rt_col_name)
}
\arguments{
\item{gc_peak_list}{List of data.frames. Each data.frame contains GC-data (e.g. retention time, peak area, peak height) of one sample. Variables are stored in columns.
Rows represent distinct peaks. Retention time is a required variable.}

\item{reference}{a character string indicating a sample included in \code{gc_peak_list} used as a reference to align to.}

\item{max_linear_shift}{defines a window to search for an optimal linear shift of samples
with respect to the reference. Shifts are evaluated within - \code{max_linear_shift to + max_linear_shift}.
Default is 0.05.}

\item{step_size}{indicates the step size in which linear shifts are evaluated
between \code{max_linear_shift} and \code{-max_linear_shift}.}

\item{error}{numeric value defining the allowed difference in retention times in
derterming if two peaks are shared. The default \code{error=0} counts
a peak a shared if retention times match excatly.}

\item{rt_col_name}{character string naming the column holding retention times.The designated variable needs to be numeric.}
}
\value{
A list of data.frames containing chromatograms with applied linear shifts
}
\description{
\code{linear_transformation()} applies small linear shifts of all peaks of
individual samples with respect to one reference. Thereby the number of shared
compounds among samples is maximized. The interval in which linear transformations are evaluated
is adjustable as well as the step size within this range.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

