% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_append.R
\name{matrix_append}
\alias{matrix_append}
\title{Adds zeros to matrices to make row numbers equal}
\usage{
matrix_append(Matrix, Chromatograms)
}
\arguments{
\item{Matrix}{A matrix to append}

\item{Chromatograms}{List of all matrices}
}
\description{
Adds zeros to matrices to make row numbers equal
}

