% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_append.R
\name{matrix_append}
\alias{matrix_append}
\title{Append zeros to matrices}
\usage{
matrix_append(gc_peak_df, gc_peak_list)
}
\arguments{
\item{gc_peak_df}{data.frame containing GC-data (e.g. retention time, peak area, peak height) of one sample. Variables are stored in columns.
Rows represent distinct peaks. Retention time is a required variable.}

\item{gc_peak_list}{List of data.frames. Each data.frame contains GC-data (e.g. retention time, peak area, peak height) of one sample. Variables are stored in columns.
Rows represent distinct peaks. Retention time is a required variable.}
}
\value{
a appended matrix.
}
\description{
\code{matrix_append()} adds zeros to a a data.frames grouped within a list.
If required rows containing just zeros are appended to match the dimensions of the
largest matrix having the most rows of the list \code{gc_peak_list}.
}
\details{
If the matrix has the same dimensions as the largest of \code{gc_peak_list}
         the input matrix is returned.
}
\keyword{internal}

