% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_chroma.R
\name{mean_of_samples}
\alias{mean_of_samples}
\title{Calculates mean of retention times up to a certain sample (Zeros are ignored)}
\usage{
mean_of_samples(chromatograms, samples, retention_row, rt_col_name)
}
\arguments{
\item{chromatograms}{\code{data.frame} containing GC data (retention times, peak area, peak hight etc) for
one individual in adjacent columns.}

\item{samples}{indices of samples up to sample of interest (1:sample-1)}

\item{retention_row}{current retention time row to be compared}

\item{rt_col_name}{name of retention time column}
}
\value{
mean of rts
}
\description{
Calculates mean of retention times up to a certain sample (Zeros are ignored)
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

