% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_retention_times.R
\name{mean_retention_times}
\alias{mean_retention_times}
\title{calculates mean retention time per row}
\usage{
mean_retention_times(gc_peak_list, rt_col_name)
}
\arguments{
\item{rt_col_name}{character string naming the column holding retention times.The designated variable needs to be numeric.}
}
\value{
vector with mean retention times per peak in \code{gc_peak_list}.
}
\description{
\code{mean_retention_times} obtains the mean retention time for every substance currently listed in
data.frames of \code{gc_peak_list}. Calculations are performed by \code{\link{mean_retention_time_row}}
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

