% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_redundant_peaks.R
\name{merge_redundant_peaks}
\alias{merge_redundant_peaks}
\title{Merges redundant rows}
\usage{
merge_redundant_peaks(gc_peak_list, min_diff_peak2peak = 0.05, rt_col_name)
}
\arguments{
\item{gc_peak_list}{List of data.frames. Each data.frame contains GC-data (e.g. retention time, peak area, peak height) of one sample. Variables are stored in columns.
Rows represent distinct peaks. Retention time is a required variable.}

\item{min_diff_peak2peak}{defines the minimum difference in retention times among distinct
substances. Substances that differ less, are merged if every sample contains either one
or none of the respective compounds. Default is 0.05.}

\item{rt_col_name}{\code{character}, name of the column holding retention times (i.e. "RT")}
}
\value{
a list of data.frames for individual samples with redundant peaks merged.
}
\description{
\code{merge_redundant_peaks()} allows to merge neighbouring peaks (i.e substances) of similiar retention time. All peaks
that are similar in retention times (definded by \code{min_diff_peak2peak}) are combined, in case that
samples contain either one or none of two compared peaks, but never both. An optional procedure
allows to merge close peak if only a small subset of samples contains both of them.
}
\details{
Peaks that are potentially redundant are determined by comparing the similarity of retention times with
respect to \code{min_diff_peak2peak}. Candidates are all peaks that are closer to each other than this
threshold. After one pair of rows was merged average retention times are calculated again. The algorithm
stops if no pair of similar peaks remains or none of those appears redunant (i.e. at least one sample
shows both peaks).
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

