% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rows.R
\name{merge_redundant_rows}
\alias{merge_redundant_rows}
\title{merges redundant rows}
\usage{
merge_redundant_rows(chromatograms, average_rts, min_distance = 0.05,
  rt_col_name)
}
\arguments{
\item{chromatograms}{\code{data.frame} containing GC data (retention times, peak area, peak hight etc) for
one individual in adjacent columns. The first column for all individuals has to be the retention
time, retention time has to be named RT.}

\item{average_rts}{average retention times across samples per row}

\item{min_distance}{difference between the mean retention time of two rows of the chromatograms
to be considered for merging if no individual has substances in both rows.}
}
\value{
chromatograms with merged rows
}
\description{
merges redundant rows
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

