% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rows.R
\name{merge_rows}
\alias{merge_rows}
\title{algorithm to adjust unreliability of chromatograms on an individual peak basis}
\usage{
merge_rows(chromatogram, to_merge, criterion = "zero", rt_col_name)
}
\arguments{
\item{chromatogram}{\code{data.frame} containing GC data (retention times, peak area, peak hight etc) for
one individual in adjacent columns. The first column for all individuals has to be the retention
time, retention time has to be named RT.}

\item{to_merge}{Last row of a similar pair}

\item{criterion}{"Zero" - delete a row containing zeros, or
"area": take the larger of two peaks, defined by the are of the peaks. "area" is just
useful when redundancy criterion was "proportional" instead of the default "strict".}

\item{rt_col_name}{name of the retention time column.}
}
\value{
aligned chromatograms
}
\description{
algorithm to adjust unreliability of chromatograms on an individual peak basis
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

