% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rows.R
\name{merge_rows}
\alias{merge_rows}
\title{Merge two rows into one}
\usage{
merge_rows(gc_peak_df, to_merge, criterion = "strict", rt_col_name,
  conc_col_name)
}
\arguments{
\item{gc_peak_df}{data.frame containing GC-data (e.g. retention time, peak area, peak height) of one sample. Variables are stored in columns.
Rows represent distinct peaks. Retention time is a required variable.}

\item{to_merge}{integer vector indicating the index of the last row of a pair to merge.}

\item{criterion}{character string either "strict" or "proportional". "strict" is used to delete the row containing zeros,
"proportional" provokes the selection of the larger of two peaks, defined by the concentration of the peaks.
\code{criterion="proportional"} will only affect pairs of peaks where less than 5% of the samples containg
two conflicting peaks.}

\item{rt_col_name}{Character string naming the column containing retention times.The designated variable needs to be numeric.}

\item{conc_col_name}{Character string naming a column used for the quantification of peaks (e.g. peak area or peak height).
The designated variable needs to be numeric.}
}
\value{
aligned chromatograms
}
\description{
Merges the content of two rows in such a way that content is retained. This means that either the row
is selected that contains in case that the second is empty (i.e just zeros). If both rows have a content
the row with the larger concentration (i.e peak area or peak height) is retained in \code{gc_peak_df}
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}
\keyword{internal}

