% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_peaks.R
\name{norm_peaks}
\alias{norm_peaks}
\title{Normalisation of peaks}
\usage{
norm_peaks(GCout, conc_col_name = NULL, rt_col_name = NULL,
  out = c("list", "data.frame"))
}
\arguments{
\item{GCout}{object of class GCaling created with \link{align_chromatograms}. Contains a list
of data.frames including the retention time and other variables, of which one needs
to be named as specified by \code{conc_col_name}.}

\item{conc_col_name}{character string denoting a column in data.frames of \code{GCout}
containing a variable describing the abundance of peaks (e.g. peak area or peak height).}
}
\value{
a list of data.frames containing normalised peak abundances

 @author Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}
\description{
\code{norm_peaks} calculates the relative abundance of a peaks by normalising
with regard to the cumulative abundance of all peaks that are present within an
individual chromatogram. The desired measure of peak abundance needs to be a column
within the original gas-chromatography data \code{datafile} submitted to
\link{align_chromatograms}.
}
\keyword{internal}

