% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_shift.R
\name{peak_shift}
\alias{peak_shift}
\title{shifts peaks}
\usage{
peak_shift(gc_peak_df, ref_df, max_linear_shift = 0.05, step_size = 0.01,
  error = 0, rt_col_name)
}
\arguments{
\item{gc_peak_df}{data.frame containing GC-data (e.g. retention time, peak area, peak height) of one sample. Variables are stored in columns.
Rows represent distinct peaks. Retention time is a required variable.}

\item{ref_df}{data.frame of form \code{gc_peak_df} serving as a reference.}

\item{max_linear_shift}{Defines a window to search for an optimal linear shift of chromatogram peaks
with respect to the reference. Shifts are evaluated within - \code{max_linear_shift to + max_linear_shift}.
Default is 0.05.}

\item{step_size}{defines the increments between \code{-max_linear_shift}:\code{max_linear_shift}
to consider in the search for an optimal linear transformation of peaks.}

\item{error}{numeric value defining the allowed difference in retention times in
derterming if two peaks are shared. The default \code{error=0} counts
a peak a shared if retention times match excatly.}

\item{rt_col_name}{Character string naming the column holding retention times.The designated variable needs to be numeric.}
}
\value{
Numeric Value, indicating the best shift (e.g. -0.02 seconds)
}
\description{
shifts peaks of individual peak data.frames.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
  (meinolf.ottensmann@web.de)
}

