% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_transformation.R
\name{peak_shift}
\alias{peak_shift}
\title{shifts peaks}
\usage{
peak_shift(sample_df, ref_df, shift = 0.05, step_size = 0.01, error = 0,
  rt_col_name)
}
\arguments{
\item{sample_df}{data.frame containing individual GC data

 @param ref_df data.frame containing individual GC data of the reference chromatogram}

\item{shift}{the maximum shift in retention times that will be considered in estimating the best transformation.}

\item{step_size}{indicates the step size in which linear shifts are evaluated
between \code{shift} and \code{-shift}.}

\item{error}{numeric value defining the allowed difference in retention times in
derterming if two peaks are shared. The default \code{error=0} counts
a peak a shared if retention times match excatly.}

\item{rt_col_name}{character string denoting the column containing retention times
in data.frames of \code{chromatograms}}
}
\value{
Numeric Value, indicating the best shift (e.g. -0.02 seconds)
}
\description{
shifts peaks of individual chromatograms
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
  (meinolf.ottensmann@web.de)
}

