% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.GCalign.R
\name{print.GCalign}
\alias{print.GCalign}
\title{print.GCalign: Gives an overview about an executed alignment}
\usage{
\method{print}{GCalign}(x, ...)
}
\arguments{
\item{x}{Object of class \code{GCalign}}

\item{...}{actually not useful here, but required to pass the check (S3 method of print)}
}
\description{
print.GCalign: Gives an overview about an executed alignment
}
\details{
for function arguments that have not been entered the default are printer, but
these are update automatically, if these are changed in the main funtion!
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
 (meinolf.ottensmann@web.de)
}

