% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.GCalign.R
\name{print.GCalign}
\alias{print.GCalign}
\title{Summarizing Peak Alignments with GCalignR}
\usage{
\method{print}{GCalign}(x, write_text_file = FALSE, ...)
}
\arguments{
\item{x}{an object of class \strong{"GCalign"}, the result of a call to \code{\link{align_chromatograms}}}

\item{write_text_file}{logical. If \code{TRUE} the output is passed to \code{.txt} file. No text is printed
to the console in this case.}

\item{...}{optional arguments passed on to methods. Internally \link{cat} is used to generate output, hence optional arguments are currently not ignored.}
}
\description{
print method for class "GCalign"
}
\examples{
print(aligned_peak_data) # prints summary to the Console

}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
 (meinolf.ottensmann@web.de)
}

