% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.GCalign.R
\name{print.GCalign}
\alias{print.GCalign}
\title{Summarising Peak Alignments with GCalignR}
\usage{
\method{print}{GCalign}(x, write_text_file = FALSE, ...)
}
\arguments{
\item{x}{Object of class GCalign, created with \code{\link{align_chromatograms}}}

\item{write_text_file}{A boolean allowing to write a text file.}

\item{...}{Optional arguments passed on to methods are currently not supported.}
}
\description{
print method for class "GCalign"
}
\examples{
## GCalign object
data("aligned_peak_data")
## print summary
print(aligned_peak_data)

}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
 (meinolf.ottensmann@web.de)
}
