% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_cutoff.R
\name{rt_cutoff}
\alias{rt_cutoff}
\title{cuts peaks below and above specified retention times}
\usage{
rt_cutoff(gc_peak_df, rt_col_name, low = NULL, high = NULL)
}
\arguments{
\item{gc_peak_df}{data.frame containing GC-data (e.g. retention time, peak area, peak height) of one sample. Variables are stored in columns.
Rows represent distinct peaks. Retention time is a required variable.}

\item{rt_col_name}{character string with the name of the retention time column in \code{gc_peak_df}}

\item{low}{lower threshold for retention times. RTs higher than \code{low} will be kept}

\item{high}{upper threshold for retention times. RTs lower than \code{high} will be kept}
}
\value{
data.frame without retention times inside \code{low}:\code{high}
}
\description{
By specifying cut-off thresholds retention times above (\code{highw}) or below (\code{high}) these peaks can
             be removed from the data in order to retain a subset of peaks belonging to
             specific retention times. All cut-off values are given in minutes.
}
\details{
In addition to cutting retention times, this function removes rows of \code{gc_peak_df} containing
         purely NA. These are artefacts added by \code{\link[utils]{read.table}} when importing the data.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

