% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_extract.R
\name{rt_extract}
\alias{rt_extract}
\title{Extract retention times to plot a heatmap}
\usage{
rt_extract(gc_peak_list, rt_col_name)
}
\arguments{
\item{gc_peak_list}{List of data.frames. Each data.frame contains GC-data (e.g. retention time, peak area, peak height) of one sample. Variables are stored in columns.
Rows represent distinct peaks. Retention time is a required variable.}

\item{rt_col_name}{Character string naming the column holding retention times.The designated variable needs to be numeric.}
}
\value{
data.frame containing retention times of all samples in \code{gc_peak_list}.
         Samples are sorted in rows, columns represent substances. Columns are named
         by the mean retention time based on all samples containing a retention time
         > 0 at this location.
}
\description{
\code{rt_extract()} extracts retention times for all samples within a
             a list of chromatograms and formats them to one matrix.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}
\keyword{internal}

