% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_extract.R
\name{rt_extract}
\alias{rt_extract}
\title{Extract retention times to plot a heatmap}
\usage{
rt_extract(chromatograms, rt_name)
}
\arguments{
\item{chromatograms}{list of data.frames containing GC data for individual samples.}

\item{rt_name}{chracter string specifies column holding retention times}
}
\value{
data.frame containing retention times of all samples in \code{chromatograms}.
         Samples are sorted in rows, columns represent substances. Columns are named
         by the mean retention time based on all samples containing a retention time
         > 0 at this location.
}
\description{
\code{rt_extract()} extracts retention times for all samples within a
             a list of chromatograms and formats them to one matrix.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

