% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_transformation.R
\name{shared_peaks}
\alias{shared_peaks}
\title{Calculate the Number of shared peaks between a Chromatogram and its reference}
\usage{
shared_peaks(sample_df, ref_df, shift_steps, error = 0, rt_col_name)
}
\arguments{
\item{sample_df}{\code{data.frame} with individual gc data}

\item{ref_df}{\code{data.frame} with individual gc data of the reference chromatogram}

\item{shift_steps}{steps to taken within the shift span}

\item{error}{allowed error}

\item{rt_col_name}{RT time column name}
}
\value{
list of number of peaks and shift steps
}
\description{
Calculate the Number of shared peaks between a Chromatogram and its reference
}
\details{
Shared peaks fall within the retention time of the reference and +- the Error [s]
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
  (meinolf.ottensmann@web.de)
}

