% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_peaks.R
\name{shared_peaks}
\alias{shared_peaks}
\title{Determines the number of shared peaks between a chromatogram and a reference}
\usage{
shared_peaks(gc_peak_df, ref_df, shift_steps, error = 0, rt_col_name)
}
\arguments{
\item{gc_peak_df}{data.frame containing GC-data (e.g. retention time, peak area, peak height) of one sample. Variables are stored in columns.
Rows represent distinct peaks. Retention time is a required variable.}

\item{ref_df}{data.frame of form \code{gc_peak_df} serving as a reference.}

\item{shift_steps}{numeric vector of shifts for which the number of shared peaks is determined.}

\item{error}{numeric value defining the allowed difference in retention times in
of two shared peaks. The default \code{error=0} counts
a peak as shared if retention times match excatly.}

\item{rt_col_name}{Character string naming the column holding retention times.The designated variable needs to be numeric.}
}
\value{
a data.frame containing the number of shared peaks and corresponding shifts for
all evaluated steps
}
\description{
Estimate the most suitable linear shift in seconds to maximize the number of
shared peaks with the reference
}
\details{
Shared peaks fall within the retention time of the reference and +- the Error (denoted in seconds)
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
  (meinolf.ottensmann@web.de)
}

