% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rows.R
\name{similar_rows}
\alias{similar_rows}
\title{Estimate degree of similarity between subsequent rows by comparison of mean retention times of rows.}
\usage{
similar_rows(average_rts, min_distance = 0.05)
}
\arguments{
\item{average_rts}{average retention times per row across individuals}

\item{min_distance}{default 0.05,}
}
\value{
position of a row, that is similar to the previous, i.e. row 5 is similar to 4
}
\description{
Estimate degree of similarity between subsequent rows by comparison of mean retention times of rows.
}
\details{
Similarity is evaluated at the level of min_distance, given in seconds
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

