% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_chroma.R
\name{simple_chroma}
\alias{simple_chroma}
\title{Simulating a simple chromatogram}
\usage{
simple_chroma(peaks = c(10, 13, 25, 37, 50))
}
\arguments{
\item{peaks}{vector of mean retention times representing peaks in a chromatogram}
}
\description{
Simulating a simple chromatogram
}
\keyword{internal}
