% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.GCalign.R
\name{summary.GCalign}
\alias{summary.GCalign}
\title{print.GCalign: Prints a the protocol of an alignment process exectuted via align_chromatogtams}
\usage{
\method{summary}{GCalign}(object, ...)
}
\arguments{
\item{object}{Object of class \code{GCalign}}
}
\description{
print.GCalign: Prints a the protocol of an alignment process exectuted via align_chromatogtams
}
\details{
Prints the LogFile written during the alignment-process, if one was created.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) & Meinolf Ottensmann
 (meinolf.ottensmann@web.de)

 @export
}

