% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_chroma.R
\name{var_of_samples}
\alias{var_of_samples}
\title{calculates mean of retention times up to a certain sample (not including 0s)}
\usage{
var_of_samples(chromatograms, samples, retention_row, rt_col_name)
}
\arguments{
\item{chromatograms}{\code{data.frame} containing GC data (retention times, peak area, peak hight etc) for
one individual in adjacent columns. The first column for all individuals has to be the retention
time, retention time has to be named RT.}

\item{samples}{indices of samples up to sample of interest (1:sample-1)}

\item{retention_row}{\code{integer} defining the current retention time row to be evaluated}

\item{rt_col_name}{\code{character} name of retention time column}
}
\value{
var of rts
}
\description{
calculates mean of retention times up to a certain sample (not including 0s)
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com) &
        Meinolf Ottensmann (meinolf.ottensmann@web.de)
}

